Removed "start button" function
Changes to index.html and main.js. This "defer" change allows the page to load images on page immediately. Slideshow is therefore ready to go on load. (w3: If the defer attribute is set, it specifies that the script is downloaded in parallel to parsing the page, and executed after the page has finished parsing)